+++ packages/fetcher/src/factory/index.ts
@@ -6,8 +6,9 @@ import {
   TArgsForPost
 } from '../types';
 import Fetcher from '../fetcher';
-import requestWithNoBody from '../util/request-with-no-body';
-import requestWithBody from '../util/request-with-body';
+
+import requestWithNoBody from './request-with-no-body';
+import requestWithBody from './request-with-body';
 
 /**
  * 这里会创建 Fetcher 实例，但不会直接把实例返回，因为那样的话用起来会不舒服（方法无法脱离实例进行调用），