+++ packages/fetcher/src/factory/request-with-no-body.ts
@@ -4,8 +4,10 @@ import {
   TArgsForGet,
   TFetcherOptionsForQuickFn
 } from '../types';
-import mergeConfig from '../util/merge-config';
 import Fetcher from '../fetcher';
+import {
+  mergeConfig
+} from '../util';
 
 /**
  * 用于执行不带 body 的请求，对应点 method 有 'GET' / 'DELETE' / 'HEAD' / 'OPTIONS' / 'JSONP'