+++ packages/fetcher/src/interceptors/request/last.ts
@@ -1,9 +1,11 @@
 import {
   IFetcherConfig
 } from '../../types';
-import isCors from '../../util/is-cors';
-import isJsonp from '../../util/is-jsonp';
-import canHaveBody from '../../util/can-have-body';
+import {
+  isCors,
+  isJsonp,
+  canHaveBody
+} from '../../util';
 
 /**
  * request 最后一个拦截器，写入 _timeStarted，如果不是 JSONP，对 headers 和 credentials 做补充