+++ packages/fetcher/CHANGELOG.md
@@ -1,9 +1,14 @@
 HISTORY
 ===
 
+# 1.3.0 2022/03/17 @驳是
+
+* FEAT 新增 helper `preventEscCancelling`
+* FIX `util` 有循环依赖
+
 # 1.2.3 2022/03/17 @驳是
 
-* 支持全 method（包括 JSONP）的 abort
+* FEAT 支持全 method（包括 JSONP）的 abort
 
 # 1.0.0 2020/11/30 @驳是